Refactor where serialization happens
This allows overriding behavior via one class.
