+++ src/Nerdbank.Qif/BankTransaction.cs
@@ -61,122 +61,9 @@ public partial record BankTransaction(DateTime Date, decimal Amount)
     public string? AccountName { get; init; }
 
     /// <summary>
-    /// Deserializes a <see cref="BankTransaction"/> from the given <see cref="QifReader"/>.
-    /// </summary>
-    /// <param name="reader">The reader to deserialize from.</param>
-    /// <returns>The deserialized record.</returns>
-    /// <devremarks>
-    /// Keep this in close sync with <see cref="MemorizedTransaction.Load(QifReader)"/>.
-    /// </devremarks>
-    public static BankTransaction Load(QifReader reader)
-    {
-        DateTime? date = null;
-        decimal? amount = null;
-        ClearedState clearedStatus = ClearedState.None;
-        string? number = null;
-        string? payee = null;
-        string? memo = null;
-        string? category = null;
-        ImmutableList<string> address = ImmutableList<string>.Empty;
-        ImmutableList<string> splitCategories = ImmutableList<string>.Empty;
-        ImmutableList<string> splitMemos = ImmutableList<string>.Empty;
-        ImmutableList<decimal> splitAmounts = ImmutableList<decimal>.Empty;
-        ImmutableList<decimal> splitPercentage = ImmutableList<decimal>.Empty;
-        while (reader.TryReadField(out ReadOnlyMemory<char> fieldName, out _))
-        {
-            if (QifUtilities.Equals(FieldNames.Date, fieldName))
-            {
-                date = reader.ReadFieldAsDate();
-            }
-            else if (QifUtilities.Equals(FieldNames.Amount, fieldName))
-            {
-                amount = reader.ReadFieldAsDecimal();
-            }
-            else if (QifUtilities.Equals(FieldNames.ClearedStatus, fieldName))
-            {
-                clearedStatus = reader.ReadFieldAsClearedState();
-            }
-            else if (QifUtilities.Equals(FieldNames.Number, fieldName))
-            {
-                number = reader.ReadFieldAsString();
-            }
-            else if (QifUtilities.Equals(FieldNames.Payee, fieldName))
-            {
-                payee = reader.ReadFieldAsString();
-            }
-            else if (QifUtilities.Equals(FieldNames.Memo, fieldName))
-            {
-                memo = reader.ReadFieldAsString();
-            }
-            else if (QifUtilities.Equals(FieldNames.Category, fieldName))
-            {
-                category = reader.ReadFieldAsString();
-            }
-            else if (QifUtilities.Equals(FieldNames.Address, fieldName))
-            {
-                address = address.Add(reader.ReadFieldAsString());
-            }
-            else if (QifUtilities.Equals(FieldNames.SplitCategory, fieldName))
-            {
-                splitCategories = splitCategories.Add(reader.ReadFieldAsString());
-            }
-            else if (QifUtilities.Equals(FieldNames.SplitMemo, fieldName))
-            {
-                splitMemos = splitMemos.Add(reader.ReadFieldAsString());
-            }
-            else if (QifUtilities.Equals(FieldNames.SplitAmount, fieldName))
-            {
-                splitAmounts = splitAmounts.Add(reader.ReadFieldAsDecimal());
-            }
-            else if (QifUtilities.Equals(FieldNames.SplitPercent, fieldName))
-            {
-                splitPercentage = splitPercentage.Add(reader.ReadFieldAsDecimal());
-            }
-        }
-
-        if (splitCategories.Count != splitMemos.Count ||
-            splitCategories.Count != Math.Max(splitAmounts.Count, splitPercentage.Count))
-        {
-            throw new InvalidTransactionException("Inconsistent number of fields for splits.");
-        }
-
-        ImmutableList<BankSplit> splits = ImmutableList<BankSplit>.Empty;
-        if (splitCategories.Count > 0)
-        {
-            var splitsBuilder = splits.ToBuilder();
-            for (int i = 0; i < splitCategories.Count; i++)
-            {
-                BankSplit split = new(splitCategories[i], splitMemos[i])
-                {
-                    Amount = splitAmounts.Count > i ? splitAmounts[i] : null,
-                    Percentage = splitPercentage.Count > i ? splitPercentage[i] : null,
-                };
-                splitsBuilder.Add(split);
-            }
-
-            splits = splitsBuilder.ToImmutable();
-        }
-
-        reader.ReadEndOfRecord();
-
-        return new(
-            ValueOrThrow(date, FieldNames.Date),
-            ValueOrThrow(amount, FieldNames.Amount))
-        {
-            ClearedStatus = clearedStatus,
-            Number = number,
-            Payee = payee,
-            Memo = memo,
-            Category = category,
-            Address = address,
-            Splits = splits,
-        };
-    }
-
-    /// <summary>
     /// The names of each field that may appear in this record.
     /// </summary>
-    private protected class FieldNames
+    internal class FieldNames
     {
         internal const string Date = "D";
         internal const string Amount = "T";
