+++ src/Nerdbank.Qif/InvestmentTransaction.cs
@@ -67,99 +67,9 @@ public record InvestmentTransaction(DateTime Date)
     public decimal AmountTransferred { get; init; }
 
     /// <summary>
-    /// Deserializes a <see cref="BankTransaction"/> from the given <see cref="QifReader"/>.
-    /// </summary>
-    /// <param name="reader">The reader to deserialize from.</param>
-    /// <returns>The deserialized record.</returns>
-    public static InvestmentTransaction Load(QifReader reader)
-    {
-        DateTime? date = null;
-        ClearedState clearedStatus = ClearedState.None;
-        string? memo = null;
-        string? action = null;
-        decimal commission = 0;
-        decimal price = 0;
-        decimal quantity = 0;
-        string? security = null;
-        string? payee = null;
-        decimal transactionAmount = 0;
-        string? accountForTransfer = null;
-        decimal amountTransferred = 0;
-        while (reader.TryReadField(out ReadOnlyMemory<char> fieldName, out _))
-        {
-            if (QifUtilities.Equals(FieldNames.Date, fieldName))
-            {
-                date = reader.ReadFieldAsDate();
-            }
-            else if (QifUtilities.Equals(FieldNames.ClearedStatus, fieldName))
-            {
-                clearedStatus = reader.ReadFieldAsClearedState();
-            }
-            else if (QifUtilities.Equals(FieldNames.Memo, fieldName))
-            {
-                memo = reader.ReadFieldAsString();
-            }
-            else if (QifUtilities.Equals(FieldNames.Action, fieldName))
-            {
-                action = reader.ReadFieldAsString();
-            }
-            else if (QifUtilities.Equals(FieldNames.Commission, fieldName))
-            {
-                commission = reader.ReadFieldAsDecimal();
-            }
-            else if (QifUtilities.Equals(FieldNames.Price, fieldName))
-            {
-                price = reader.ReadFieldAsDecimal();
-            }
-            else if (QifUtilities.Equals(FieldNames.Quantity, fieldName))
-            {
-                quantity = reader.ReadFieldAsDecimal();
-            }
-            else if (QifUtilities.Equals(FieldNames.Security, fieldName))
-            {
-                security = reader.ReadFieldAsString();
-            }
-            else if (QifUtilities.Equals(FieldNames.Payee, fieldName))
-            {
-                payee = reader.ReadFieldAsString();
-            }
-            else if (QifUtilities.Equals(FieldNames.TransactionAmount, fieldName))
-            {
-                transactionAmount = reader.ReadFieldAsDecimal();
-            }
-            else if (QifUtilities.Equals(FieldNames.AccountForTransfer, fieldName))
-            {
-                accountForTransfer = reader.ReadFieldAsString();
-            }
-            else if (QifUtilities.Equals(FieldNames.AmountTransferred, fieldName))
-            {
-                amountTransferred = reader.ReadFieldAsDecimal();
-            }
-        }
-
-        reader.ReadEndOfRecord();
-
-        return new(
-            ValueOrThrow(date, FieldNames.Date))
-        {
-            ClearedStatus = clearedStatus,
-            Memo = memo,
-            Action = action,
-            Commission = commission,
-            Price = price,
-            Quantity = quantity,
-            Security = security,
-            Payee = payee,
-            TransactionAmount = transactionAmount,
-            AccountForTransfer = accountForTransfer,
-            AmountTransferred = amountTransferred,
-        };
-    }
-
-    /// <summary>
     /// The names of each field that may appear in this record.
     /// </summary>
-    private static class FieldNames
+    internal static class FieldNames
     {
         internal const string Date = "D";
         internal const string Action = "N";
